added font awesome to nav bar, added a footer and components login and sign up
