+++ client/src/App.js
@@ -3,10 +3,13 @@ import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import GamePage from "./pages/game";
 import SignUpPage from "./pages/signup";
 import LoginPage from "./pages/login";
+import Navbar from "./components/Navbar";
+import Footer from "./components/Footer"
 
 function App() {
   return (
     <Router>
+      <Navbar></Navbar>
       <div>
         <Switch>
           <Route exact path="/signup" component={SignUpPage} /> 
@@ -14,6 +17,7 @@ function App() {
           <Route exact path="/game" component={GamePage} />
           </Switch>
       </div>
+      <Footer></Footer>
     </Router>
   );
 }
