+++ client/src/components/Footer/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+import "./style.css";
+// Depending on the current path, this component sets the "active" class on the appropriate navigation link item
+function Footer() {
+  return (
+    <footer>copright@typing tutor</footer>
+  );
+}
+
+export default Footer;
