+++ client/src/pages/login.js
@@ -1,40 +1,12 @@
-import React from 'react';
-import Axios from 'axios';
+import React, { Component } from "react";
+import Login from "../components/Login";
 
-class Login extends React.Component {
-state = {username: "", password: ""}
-handleUsername = (event) =>{
-  // console.log("typing username", event.target.value)
-  this.setState({username: event.target.value})
-}
-handlePassword = (event) =>{
-  // console.log("typing username", event.target.value)
-  this.setState({password: event.target.value})
-}
-handleSubmit = () =>{
-  console.log("about to login", this.state)
-  Axios.post("/login", {
-    username: this.state.username,
-    password: this.state.password,
-  })
-  .then(function(data){
-    console.log("from the back end", data)
-  }) 
-}
-
+class LoginPage extends Component {
   render() {
-    console.log("state", this.state)
     return (
-      <div>
-        <h1>Login</h1>
-        <p>Username</p>
-        <input onChange= {this.handleUsername}></input>
-        <p>Password</p>
-        <input onChange= {this.handlePassword}></input>
-        <button onClick= {this.handleSubmit}>Submit</button>
-      </div>
+      <Login/>
     );
   }
 }
 
-export default Login;
+export default LoginPage;
