+++ package.json
@@ -29,11 +29,13 @@
     "connect-mongo": "^3.2.0",
     "express": "^4.16.3",
     "express-session": "^1.17.0",
+    "font-awesome": "^4.7.0",
     "if-env": "^1.0.4",
     "json-schema-traverse": "^0.4.1",
     "mongoose": "^5.8.2",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
+    "react-icons": "^3.8.0",
     "react-router-dom": "^5.1.2",
     "uifx": "^2.0.7"
   }
